feat(molecule/pagination): better demo & documentation
