+++ components/molecule/pagination/src/helpers/pagination/index.js
@@ -1,0 +1,6 @@
+import processPage from './processPage'
+import prevPage from './prevPage'
+import nextPage from './nextPage'
+import range from './range'
+
+export default {processPage, prevPage, nextPage, range}
