+++ demo/molecule/pagination/playground
@@ -66,6 +66,9 @@ return (
     <div style={fieldStyles}>
       <h3>Extended Version</h3>
       <DynamicMoleculePagination totalPages={25} page={17} />
+
+      <h3>Extended Version <code>showPages=7</code></h3>
+      <DynamicMoleculePagination totalPages={25} page={17} showPages={7}/>
     </div>
     <div style={fieldStyles}>
       <h3>Compressed Version</h3>
@@ -74,6 +77,13 @@ return (
     {/* ------------------------------------------------------------------------------------------------------------- */}
     <h2>Static</h2>
     <h3>Extended Version</h3>
+    <div style={fieldStyles}>
+      <h4>Basic</h4>
+      <p>
+        <code>totalPages=25 page=7</code>
+      </p>
+      <MoleculePagination totalPages={25} page={7} />
+    </div>
     <div style={fieldStyles}>
       <h4>w/ Next</h4>
       <p>
